Make memory_manager a subclass of the new resource_manager
