+++ tpie/CMakeLists.txt
@@ -107,6 +107,7 @@ set (HEADERS
 		progress_indicator_null.h
 		progress_indicator_terminal.h
 		queue.h
+		resource_manager.h
 		serialization.h
 		serialization2.h
 		serialization_stream.h
@@ -174,6 +175,7 @@ set (SOURCES
 	prime.cpp
 	progress_indicator_base.cpp
 	progress_indicator_subindicator.cpp
+	resource_manager.cpp
 	serialization_stream.cpp
 	hash.cpp
 	tempname.cpp
